Cleanup regex
Special characters lose their special meaning in character sets, so
there's no need to escape them with a backslash. Furthermore,
some backslash escaped characters have no special meaning,
so the backslash can be removed as well.

"]" does normally close a set, except when it's the first character.
Similarly, "-" has no special meaning at the beginning or the end.
Therefore, the order in a few character sets had to be slightly
modified.

Just to be sure, the old and new versions of the regexp were confirmed
to be equivalent via re.compile with the re.DEBUG flag and comparing the
output.