Corrected Typo in Portuguese README
I fixed a minor typo in the Portuguese version of the README file. The word "Markdowm" should be "Markdown". This improves the clarity of the document. I am searching for another typo, but i can't find other typo, this is good 😃.  
I have a issue asking for fix the typo [here](https://github.com/Textualize/rich/issues/3172).