Drop a test that has not been used since it was committed
This test has been dead code since its introduction and does not get
executed by pytest due to starting with `text` rather than `test`. The
underlying data for this test is also non-trivial to update and
dependent on the specific Python version.

Since this is already dead code and is non-trivial to update, removing
this test more clearly reflects what happens with this dead code in
tests and prevents potential confusion due to its presence.